why reson

+++ drawkorchart.py
@@ -275,7 +275,7 @@ def cash_flow(company_name, cf_an, cf_qu, in_df):
             #fig.update_traces(texttemplate='%{text:.3s}') 
             fig.update_yaxes(title_text='현금흐름', range=[0, max(cf_an.loc[:,y_data_bar[0]])*2], secondary_y = False)
             #fig.update_yaxes(title_text='Profit', range=[0, max(income_df.loc[:,y_data_bar[0]])*2], secondary_y = False)
-            fig.update_yaxes(title_text='FCFF', range=[-max(in_df.loc[:,y_data_line[0].astype(int)]), max(in_df.loc[:,y_data_line[0]].astype(int))* 1.2], secondary_y = True)
+            fig.update_yaxes(title_text='FCFF', range=[-max(in_df.loc[:,y_data_line[0]]), max(in_df.loc[:,y_data_line[0]])* 1.2], secondary_y = True)
             fig.update_yaxes(showticklabels= True, showgrid = False, zeroline=True, ticksuffix="억원", secondary_y = False)
             fig.update_yaxes(showticklabels= True, showgrid = False, zeroline=True, ticksuffix="억원", secondary_y = True)
             fig.update_layout(title = titles, titlefont_size=15, legend=dict(orientation="h"), template=template)#, xaxis_tickformat = '%Y.%m')
